roll_tracker_backend/coin_collection/serializers.py
from rest_framework import serializers
from .models import CoinCollection


class CollectionSerializer(serializers.ModelSerializer):
    class Meta:
        model = CoinCollection
        fields = '__all__'